A whole ton of stuff:
Added the eval.dev.java.net package for equation evaluations
Added a new pluggable equation evaluator framework so that both JEP and Eval can be in place at the same time, and ~config can be used to select which one to use. A future commit might also allow the engine to be selected with the {{eval}} factoid, so that {{eval||jep||2+3}} could be used (as an example).
Added support for dynamically changing the list of superops. Basically, this implements the ~superop command.
Changed ~op to delimit the list of ops with 4 spaces instead of sending each in its own message.
Changed ~leave to allow channel operators to use it, in case they don't want the bot at their channel.
Added a configuration setting to ~config for selecting the eval engine.
Added the error and kick functions, for throwing a factoid exception and kicking a user off the channel (not implemented yet), respectively.
Added a "restricted" boolean to factoids. This will soon be accompanied by the ~restrict and ~unrestrict (and probably ~isrestricted) commands to restrict a factoid to only being called by marlen or ops. Importing a factoid with {{import}} causes it to be run as if marlen had called it; this allows for, as an example, not restricting ~roulette but restricting all factoids (such as roulette-create, roulette-reset, etc) that it uses.
Added JZUtils, a set of IRC utilities backported from JZBot 3

git-svn-id: https://jwutils.googlecode.com/svn/trunk@539 9038e2a8-f00e-11dd-89c5-9df526a60542