Reconnect immediately when cycling due to channel rename
One of these days I'll implement actual proper support for channel
renames. It's tricky, though, because it's not something IRC
supports and therefore not something I even considered when
designing JZBot. Probably won't do anything special, and will just
emulate a part from the old channel and a join to the new.